tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-embed-src.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="./support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="require-trusted-types">
</head>
<body>
<script>
  // String assignments throw.
  test(t => {
    var s = document.createElement('embed');
    assert_throws(new TypeError(), _ => {
      s.src = "Fail.";
    });
    assert_equals('', s.src);
  }, "src = 'string' assignment throws.");

  // TrustedURL assignments throw.
  test(t => {
    var url = TrustedURL.unsafelyCreate(URLS.safe);

    var s = document.createElement('embed');
    assert_throws(new TypeError(), _ => {
      s.src = url;
    });
    assert_equals('', s.src);
  }, "src = TrustedURL.unsafelyCreate(URLS.safe) assignment throws");

  // TrustedScriptURL assignments work.
  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.safe);

    var s = document.createElement('embed');
    s.src = url;
    assert_equals(url + '', s.src);
  }, "src = TrustedScriptURL.unsafelyCreate(URLS.safe)");

  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.javascript);

    var s = document.createElement('embed');
    s.src = url;
    assert_equals(url + '', s.src);
  }, "src = TrustedScriptURL.unsafelyCreate(URLS.javascript)");

  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.external);

    var s = document.createElement('embed');
    s.src = url;
    assert_equals(url + '', s.src);
  }, "src = TrustedScriptURL.unsafelyCreate(URLS.external)");
</script>
